added pages and changed layout
