+++ main_start_page.py
@@ -4,22 +4,32 @@ from streamlit_extras.app_logo import add_logo
 from streamlit_extras.colored_header import colored_header
 
 
+st.set_page_config( page_title="Wavecraft", page_icon=":musical_note:")
 
 
-def register_page():
-    st.title("Register Page")
-    uploaded_file = st.file_uploader("Upload a wav-file or a folder")
 
-def recognise_page():
-    st.title("Recognise Page")
-    uploaded_file = st.file_uploader("Upload a wav-file or a folder")
-    if st.button("Activate Microphone"):
+show_pages(
+    [
+        Page("register_page.py", "Upload your songs", "🎵"),
+        Page("recognise_page.py", "Find the song", "🔮")
+    ]
+)
+
+
+#def register_page():
+    #st.title("Register Page")
+    #uploaded_file = st.file_uploader("Upload a wav-file or a folder")
+
+#def recognise_page():
+    #st.title("Recognise Page")
+    #uploaded_file = st.file_uploader("Upload a wav-file or a folder")
+    #if st.button("Activate Microphone"):
         #add recongnise code
-        st.write("Wavecraft is listening")
+        #st.write("Wavecraft is listening")
 
 def add_logo():
     logo_image = "crying_cat.jpg"
-    st.image(logo_image, use_column_width=True)
+    st.image(logo_image, width=200)
 
 
 
@@ -28,13 +38,13 @@ def main():
     add_logo()
 
     st.title("Wavecraft")
-    st.write("Welcome to Wavecraft! This website let you recognise your favorite songs and unknown musical pieces")
+    st.markdown("Welcome to Wavecraft! This website lets you recognise your favorite songs and unknown musical pieces")
 
-    if st.button("Register"):
-        register_page()
+    #if st.button("Register"):
+        #register_page()
 
-    if st.button("Recognise"):
-        recognise_page()
+    #if st.button("Recognise"):
+        #recognise_page()
 
 if __name__ == "__main__":
     main()
